requirements.txt
Django>=1.7.1
python-social-auth>=0.2.1
Pillow>=2.6.1
django-imagekit>=3.2.4
django-braces>=1.4.0
django-bootstrap-form>=3.1
